requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.1.2
chromadb==0.5.5
rank-bm25==0.2.2
cohere==5.9.0
groq==0.11.0
streamlit==1.38.0
ragas==0.1.21
sentence-transformers==3.1.0
pypdf==4.3.1
python-dotenv==1.0.1
pandas==2.2.2
datasets==2.21.0
